+++ service/controller/app/resources.go
@@ -44,6 +44,7 @@ type appResourcesConfig struct {
 	ProjectName       string
 	Provider          string
 	UniqueApp         bool
+	WorkloadClusterID string
 }
 
 func newAppResources(config appResourcesConfig) ([]resource.Interface, error) {
@@ -163,7 +164,8 @@ func newAppResources(config appResourcesConfig) ([]resource.Interface, error) {
 		c := chart.Config{
 			Logger: config.Logger,
 
-			ChartNamespace: config.ChartNamespace,
+			ChartNamespace:    config.ChartNamespace,
+			WorkloadClusterID: config.WorkloadClusterID,
 		}
 
 		ops, err := chart.New(c)
@@ -268,7 +270,8 @@ func newAppResources(config appResourcesConfig) ([]resource.Interface, error) {
 			CtrlClient: config.K8sClient.CtrlClient(),
 			Logger:     config.Logger,
 
-			ChartNamespace: config.ChartNamespace,
+			ChartNamespace:    config.ChartNamespace,
+			WorkloadClusterID: config.WorkloadClusterID,
 		}
 
 		statusResource, err = status.New(c)