+++ service/watcher/chartstatus/chartstatus.go
@@ -19,17 +19,16 @@ import (
 	"k8s.io/apimachinery/pkg/watch"
 )
 
-const chartOperatorAppName = "chart-operator"
-
 var chartResource = schema.GroupVersionResource{Group: "application.giantswarm.io", Version: "v1alpha1", Resource: "charts"}
 
 type ChartStatusWatcherConfig struct {
 	K8sClient k8sclient.Interface
 	Logger    micrologger.Logger
 
-	ChartNamespace string
-	PodNamespace   string
-	UniqueApp      bool
+	ChartNamespace    string
+	PodNamespace      string
+	UniqueApp         bool
+	WorkloadClusterID string
 }
 
 type ChartStatusWatcher struct {
@@ -37,9 +36,10 @@ type ChartStatusWatcher struct {
 	kubeConfig kubeconfig.Interface
 	logger     micrologger.Logger
 
-	appNamespace   string
-	chartNamespace string
-	uniqueApp      bool
+	chartNamespace    string
+	podNamespace      string
+	uniqueApp         bool
+	workloadClusterID string
 }
 
 func NewChartStatusWatcher(config ChartStatusWatcherConfig) (*ChartStatusWatcher, error) {
@@ -54,7 +54,7 @@ func NewChartStatusWatcher(config ChartStatusWatcherConfig) (*ChartStatusWatcher
 		return nil, microerror.Maskf(invalidConfigError, "%T.ChartNamespace must not be empty", config)
 	}
 	if config.PodNamespace == "" {
-		return nil, microerror.Maskf(invalidConfigError, "%T.PodNamespace must not be empty", config)
+		return nil, microerror.Maskf(invalidConfigError, "%T.WatchNamespace must not be empty", config)
 	}
 
 	var kubeConfig kubeconfig.Interface
@@ -76,11 +76,11 @@ func NewChartStatusWatcher(config ChartStatusWatcherConfig) (*ChartStatusWatcher
 		kubeConfig: kubeConfig,
 		logger:     config.Logger,
 
-		// We get a kubeconfig for the cluster from the chart-operator app CR
-		// which is in the same namespace as this instance of app-operator.
-		appNamespace:   config.PodNamespace,
-		chartNamespace: config.ChartNamespace,
-		uniqueApp:      config.UniqueApp,
+		// We get a kubeconfig for the cluster from the chart-operator app CR.
+		chartNamespace:    config.ChartNamespace,
+		podNamespace:      config.PodNamespace,
+		uniqueApp:         config.UniqueApp,
+		workloadClusterID: config.WorkloadClusterID,
 	}
 
 	return c, nil
@@ -146,8 +146,13 @@ func (c *ChartStatusWatcher) watchChartStatus(ctx context.Context) {
 			}
 
 			// The chart CR is always in the giantswarm namespace so the
-			// chart CR is annotated with the app CR namespace.
+			// chart CR is annotated with the app CR name and namespace.
 			appNamespace, ok := chart.Annotations[annotation.AppNamespace]
+			if !ok {
+				c.logger.Debugf(ctx, "failed to get annotation %#q for chart %#q", annotation.AppNamespace, chart.Name)
+				continue
+			}
+			appName, ok := chart.Annotations[annotation.AppName]
 			if !ok {
 				c.logger.Debugf(ctx, "failed to get annotation %#q for chart %#q", annotation.AppNamespace, chart.Name)
 				continue
@@ -155,7 +160,7 @@ func (c *ChartStatusWatcher) watchChartStatus(ctx context.Context) {
 
 			app := v1alpha1.App{}
 			err = c.k8sClient.CtrlClient().Get(ctx,
-				types.NamespacedName{Name: chart.Name, Namespace: appNamespace},
+				types.NamespacedName{Name: appName, Namespace: appNamespace},
 				&app)
 			if err != nil {
 				c.logger.Errorf(ctx, err, "failed to get app '%s/%s'", app.Namespace, app.Name)